refactor(svc-coins): update project to respect InvariantGlobalization and minor models changes
- Change MarketCapUsd property type from int? to long? for improved precision across multiple classes.
- Clean up project files by removing obsolete properties and ensuring consistency in configuration.

+++ src/SVC_Microservices/SVC_Coins/Domain/Entities/CoinsEntity.cs
@@ -54,7 +54,7 @@ public class CoinsEntity
     /// <summary>
     /// The market capitalization in USD.
     /// </summary>
-    public int? MarketCapUsd { get; set; }
+    public long? MarketCapUsd { get; set; }
 
     /// <summary>
     /// The current price in USD.
